+++ ospray/camera/PanoramicCamera.cpp
@@ -16,4 +16,13 @@ std::string PanoramicCamera::toString() const
   return "ospray::PanoramicCamera";
 }
 
+void PanoramicCamera::commit()
+{
+  Camera::commit();
+
+  ispc::PanoramicCamera_set(getIE(),
+      (OSPStereoMode)getParam<int>("stereoMode", OSP_STEREO_NONE),
+      getParam<float>("interpupillaryDistance", 0.0635f));
+}
+
 } // namespace ospray