+++ ospray/camera/PanoramicCamera.h
@@ -1,4 +1,4 @@
-// Copyright 2009-2019 Intel Corporation
+// Copyright 2009-2020 Intel Corporation
 // SPDX-License-Identifier: Apache-2.0
 
 #pragma once
@@ -7,35 +7,14 @@
 
 namespace ospray {
 
-/*! \defgroup panoramic_camera The Panoramic Camera ("panoramic")
-
-  \brief Implements a panoramic camera with latitude/longitude mapping,
-  without support for Depth of Field or Motion Blur
-
-  \ingroup ospray_supported_cameras
-
-  A simple panoramic camera. This camera type is loaded by passing
-  the type string "panoramic" to \ref ospNewCamera
-
-  The panoramic camera supports the following parameters
-  <pre>
-  vec3f(a) pos;    // camera position
-  vec3f(a) dir;    // camera direction
-  vec3f(a) up;     // up vector
-  </pre>
-
-  The functionality for a panoramic camera is implemented via the
-  \ref ospray::PanoramicCamera class.
-*/
-
-//! Implements a simple panoramic camera (see \subpage panoramic_camera)
+// Implements a panoramic camera with latitude/longitude mapping
 struct OSPRAY_SDK_INTERFACE PanoramicCamera : public Camera
 {
-  /*! \brief constructor \internal also creates the ispc-side data structure */
   PanoramicCamera();
   virtual ~PanoramicCamera() override = default;
 
   virtual std::string toString() const override;
+  virtual void commit() override;
 };
 
 } // namespace ospray